+++ src/web/tecbank/models/BankCard.cs
@@ -8,7 +8,7 @@ namespace tecbank.models{
         public int card_num {get; set;}
         public int type {get; set;}
         public int cvc {get; set;}
-        public int balance {get; set;}
+        public float balance {get; set;}
         public String account_id {get; set;}
     }
 }
